added controllers for all new services

+++ src/PortRiskMonitor.Application/Interfaces/IWeatherConditionService.cs
@@ -2,9 +2,9 @@ namespace PortRiskMonitor.Application.Interfaces;
 
 public interface IWeatherConditionService : IIndicatorScore
 {
-    ushort GetWindSpeedKnt();
-    float GetWaterLevelCm();
-    sbyte  GetTemperatureC();
-    byte   GetHumidityPercent();
+    double GetWindSpeedKnt();
+    double GetWaterLevelCm();
+    double GetTemperatureC();
+    double GetHumidityPercent();
     string GetConditionCode();
 }
